Add scoring logic for each class: USAC, Olympic, and IFSC 2025
